Add files via upload
A script I wrote to get rid of any # character and all following characters on the same line

I tend to be so comment heavy in m coding, I often like to include a comment-free version of my scripts. Erasing the comments by hand became tedious, so I decided to just script it.

Be wary !!! In Ruby, the # character is used in string literals like:
x = 5
puts "x = #{x}"

This script will mess with you strings, for example changing the above line to
puts "x = 

Also, this really is just for my own use. It'll mess with your c code (#include) and I'm sure will have different effects on many different programming languages.